rest-service/requirements.txt
grpcio==1.73.0
grpcio-tools==1.73.0
protobuf==6.31.1
fastapi==0.115.12
uvicorn==0.34.2
pybreaker==1.3.0